Ajustes finais #26
